Phase 2/tesstrain/data/eng/requirements.txt
Pillow>=6.2.1
python-bidi>=0.4
matplotlib
pandas